Added #pragma once to header
So in C++ you usually use the same header guard trick as in C
(http://en.wikipedia.org/wiki/Include_guard)...
or you can just use #pragma once, which does the same thing.

It's not an official part of the language, but every major compiler
(gcc, clang, MSVC, intel, etc.) supports it and it leaves less room
for user error.